Move user filters to filter folder
